# openapi.yaml
openapi: 3.0.3
info:
  title: Sama SAV API
  description: API multi-tenant SaaS pour la gestion de service après-vente
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Serveur de développement
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /feedback/verify-otp:
    post:
      tags: [Feedback client]
      summary: Vérifier le code OTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                otp: { type: string, example: "123456" }
      responses:
        '200':
          description: OTP valide, retourne un accessToken
  /feedback/submit:
    post:
      tags: [Feedback client]
      summary: Soumettre le feedback
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accessToken: { type: string }
                noteTechnicien: { type: integer, minimum: 1, maximum: 5 }
                commentaireTechnicien: { type: string }
                noteEntreprise: { type: integer, minimum: 1, maximum: 5 }
                commentaireEntreprise: { type: string }
      responses:
        '200':
          description: Feedback enregistré